move template_mixin -> components
